Fixed build and introduce Rust Criterion Beanchmark tests...
The commit also include an attempt at infrastructure as code. Since, I am thinking of working
towards modifying the tooling to exploit Dapr, with podman to create the container image (deploiment maybe to docker hub or if possible using GitHub), with publishing of the solution to a K8S Dapr based deployment.